route naming and avgRating type change

+++ src/main/java/com/ssst/homefinderbackend/data/entity/RealEstateEntity.java
@@ -44,7 +44,7 @@ public class RealEstateEntity {
     private String address;
 
     @Column(name = "avg_rating", nullable = false)
-    private Integer avgRating;
+    private Double avgRating;
 
     @CreationTimestamp
     @Column(name = "created_at", updatable = false)
